LoggingConnectObserver implementation
GoCQL lacks instrumentation of its connection pool, making it difficult to determine when things have gone wrong.  Use of this observer adds a
bit of visibility by logging a message at level `ERROR` on connection failure.  Successful connections are logged at `DEBUG`.

This is being added to servicelib-golang because it probably makes
sense for every service.

Change-Id: I9fc61de0613836a8450d2d8fd866642c336e8953